+++ packages/db/src/provider.ts
@@ -1,4 +1,3 @@
-//import { toSnake } from 'snake-camel';
 import isNil from 'lodash/isNil';
 import isValidHostname from 'is-valid-hostname';
 import omit from 'lodash/omit';
